channels/JourneyTelegramBot/requirements.txt
python-telegram-bot
python-dotenv
kafka-python